added sort settings, change wp icons, new drawer animation
