+++ app/src/main/java/com/taleckij_anton/taleckijapp/launcher/recycler_training/LauncherItemHolder.java
@@ -1,4 +1,4 @@
-package com.taleckij_anton.taleckijapp.launcher;
+package com.taleckij_anton.taleckijapp.launcher.recycler_training;
 
 import android.support.annotation.ColorInt;
 import android.support.v7.widget.RecyclerView;
