+++ app/src/main/java/com/taleckij_anton/taleckijapp/launcher/recycler_training/LauncherRecyclerFragment.java
@@ -1,4 +1,4 @@
-package com.taleckij_anton.taleckijapp.launcher;
+package com.taleckij_anton.taleckijapp.launcher.recycler_training;
 
 import android.app.Fragment;
 import android.content.Context;
@@ -9,8 +9,6 @@ import android.preference.PreferenceManager;
 import android.support.annotation.ColorInt;
 import android.support.annotation.Nullable;
 import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v7.widget.AppCompatImageView;
 import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
@@ -19,9 +17,6 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ImageView;
-import android.widget.TextView;
-import android.widget.Toast;
 
 import com.taleckij_anton.taleckijapp.R;
