+++ app/src/main/res/values/arrays.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string-array name="sorttype_apps_pref_entries">
+        <item>Without Sort</item>
+        <item>By Names A-Z</item>
+        <item>By Names Z-A</item>
+        <item>Install Moment</item>
+        <item>Launch Count</item>
+    </string-array>
+
+    <string-array name="sorttype_apps_pref_entryvalues">
+        <item>0</item>
+        <item>1</item>
+        <item>2</item>
+        <item>3</item>
+        <item>4</item>
+    </string-array>
+</resources>
