+++ app/src/main/res/values/strings.xml
@@ -34,4 +34,8 @@
     <string name="launch_wp_pref_key">launch_wp_pref_key</string>
     <string name="launch_wp_pref_title">Launch Welcome Page</string>
     <string name="launch_wp_pref_summ">Launch the welcome page when the app next run(only once)</string>
+
+    <string name="sorttype_apps_pref_key">sorttype_apps_pref_key</string>
+    <string name="sorttype_apps_pref_title">Applications order</string>
+    <string name="sorttype_apps_pref_summ">Select one of the available sort types for the order of applications in your launcher</string>
 </resources>
